refactoring: set target_path from parameters, if not set in the constructor. Init temp dir in processor modules.
