+++ sinner/Core.py
@@ -10,7 +10,7 @@ from sinner.Status import Status, Mood
 from sinner.handlers.frame.BaseFrameHandler import BaseFrameHandler
 from sinner.processors.frame.BaseFrameProcessor import BaseFrameProcessor
 from sinner.typing import Frame
-from sinner.utilities import delete_subdirectories, list_class_descendants, resolve_relative_path, get_app_dir, TEMP_DIRECTORY
+from sinner.utilities import delete_subdirectories, list_class_descendants, resolve_relative_path
 from sinner.validators.AttributeLoader import Rules
 
 # single thread doubles cuda performance - needs to be set before torch import
@@ -67,11 +67,6 @@ class Core(Status):
                 'help': 'Select the frame processor from available processors'
             },
             {
-                'parameter': 'temp-dir',
-                'default': lambda: self.suggest_temp_dir(),
-                'help': 'Select the directory for temporary files'
-            },
-            {
                 'parameter': 'keep-frames',
                 'default': False,
                 'help': 'Keep temporary frames'
@@ -89,7 +84,7 @@ class Core(Status):
         current_target_path = self.target_path
         temp_resources: List[str] = []  # list of temporary created resources
         for processor_name in self.frame_processor:
-            current_processor = BaseFrameProcessor.create(processor_name, self.parameters, target_path=current_target_path, temp_dir=self.temp_dir)
+            current_processor = BaseFrameProcessor.create(processor_name, self.parameters, target_path=current_target_path)
             if current_processor.state.is_finished:
                 self.update_status(f'Processing with {current_processor.state.processor_name} already done ({current_processor.state.processed_frames_count}/{current_processor.state.frames_count})')
             else:
@@ -104,9 +99,6 @@ class Core(Status):
         if self.keep_frames is False:  # todo: add a final result check before deleting (keep frames if something wrong)
             self.update_status('Deleting temp resources')
             delete_subdirectories(self.temp_dir, temp_resources)
-
-    def suggest_temp_dir(self) -> str:
-        return self.temp_dir if self.temp_dir is not None else os.path.join(get_app_dir(), TEMP_DIRECTORY)
 
     #  returns list of all processed frames, starting from the original
     def get_frame(self, frame_number: int = 0, extractor_handler: BaseFrameHandler | None = None, processed: bool = False) -> List[Tuple[Frame, str]]:
@@ -125,7 +117,7 @@ class Core(Status):
                     self.frame_processor.remove('ResultProcessor')
                 for processor_name in self.frame_processor:
                     if processor_name not in self.preview_processors:
-                        self.preview_processors[processor_name] = BaseFrameProcessor.create(processor_name, self.parameters, target_path=self.target_path, temp_dir=self.temp_dir)
+                        self.preview_processors[processor_name] = BaseFrameProcessor.create(processor_name, self.parameters, target_path=self.target_path)
                     self.preview_processors[processor_name].load(self.parameters)
                     frame = self.preview_processors[processor_name].process_frame(frame)
                     result.append((frame, processor_name))
