+++ sinner/processors/frame/FaceSwapper.py
@@ -94,10 +94,10 @@ class FaceSwapper(BaseFrameProcessor):
             self._face_swapper = insightface.model_zoo.get_model(get_app_dir('models/inswapper_128.onnx'), providers=self.execution_providers)
         return self._face_swapper
 
-    def __init__(self, parameters: Namespace, target_path: str, temp_dir: str) -> None:
+    def __init__(self, parameters: Namespace, target_path: str | None = None) -> None:
         download_directory_path = get_app_dir('models')
         conditional_download(download_directory_path, ['https://huggingface.co/henryruhs/roop/resolve/main/inswapper_128.onnx'])
-        super().__init__(parameters, target_path, temp_dir)
+        super().__init__(parameters, target_path)
 
     def process_frame(self, frame: Frame) -> Frame:
         if self.many_faces:
